+++ packages/server/src/builder.ts
@@ -1,11 +1,13 @@
 import type { ANY_CONTRACT_PROCEDURE, ContractRouter, HTTPPath, RouteOptions, Schema, SchemaInput, SchemaOutput } from '@orpc/contract'
 import type { DecoratedLazy } from './lazy-decorated'
+import type { Middleware } from './middleware'
+import type { DecoratedMiddleware } from './middleware-decorated'
 import type { Router } from './router'
 import type { AdaptedRouter } from './router-builder'
 import type { Context, MergeContext, WELL_CONTEXT } from './types'
 import { ContractProcedure } from '@orpc/contract'
 import { type ChainableImplementer, createChainableImplementer } from './implementer-chainable'
-import { type DecoratedMiddleware, decorateMiddleware, type Middleware } from './middleware'
+import { decorateMiddleware } from './middleware-decorated'
 import { Procedure, type ProcedureFunc } from './procedure'
 import { ProcedureBuilder } from './procedure-builder'
 import { type DecoratedProcedure, decorateProcedure } from './procedure-decorated'
