+++ packages/server/src/index.ts
@@ -6,6 +6,7 @@ export * from './implementer-chainable'
 export * from './lazy'
 export * from './lazy-decorated'
 export * from './middleware'
+export * from './middleware-decorated'
 export * from './procedure'
 export * from './procedure-builder'
 export * from './procedure-caller'
